Initial CSS Grid setup and layout
This lays out the ground work for a responsive website while avoiding
micromanaging the layout as much as possible. This is achieved by using
CSS grids, which essential allows me to label each section then define
my layout with the layout names, and let the browser do all the work of
positioning for me.

Other misc changes include:
- Using more html 5 elements and trying to simplify the html code.
- Inline the Soical SVGs in the footer
- Added the ground work for the Social Follow me side bar, this includes
  the SVG's seen in the footer with the name next to the icon

To improve on:
- Broken view when reading an article, as content goes off screen,
  causing users to scroll horizontally
- Unify the header-nav bar so there is no gap
- Make the navigation bar "responsive"
- Follow me section